Split the BuildPlugin in half.
Shared global settings go into the base plugin which is used by test
libraries that wouldn't otherwise have all the extra baggage of regular
projects.
