+++ buildSrc/src/main/resources/META-INF/gradle-plugins/es.hadoop.build.base.properties
@@ -1,0 +1,1 @@
+implementation-class=org.elasticsearch.hadoop.gradle.BaseBuildPlugin
