+++ test/fixtures/minikdc/build.gradle
@@ -17,6 +17,7 @@
  * under the License.
  */
 
+apply plugin: 'es.hadoop.build.base'
 apply plugin: 'java'
 
 repositories {
