+++ test/shared/build.gradle
@@ -17,6 +17,7 @@
  * under the License.
  */
 
+apply plugin: 'es.hadoop.build.base'
 apply plugin: 'java'
 
 repositories {
@@ -31,9 +32,9 @@ java {
 dependencies {
     implementation(project(':elasticsearch-hadoop-mr'))
     
-    implementation("org.apache.hadoop:hadoop-client:${project.ext.hadoop22Version}")
-    implementation("org.apache.hadoop:hadoop-common:${project.ext.hadoop22Version}")
-    implementation("org.apache.hadoop:hadoop-mapreduce-client-core:${project.ext.hadoop22Version}")
+    implementation(project.ext.hadoopClient)
+    implementation("org.apache.hadoop:hadoop-common:${project.ext.hadoopVersion}")
+    implementation("org.apache.hadoop:hadoop-mapreduce-client-core:${project.ext.hadoopVersion}")
 
     implementation("junit:junit:${project.ext.junitVersion}")
     implementation("org.hamcrest:hamcrest-all:${project.ext.hamcrestVersion}")
